backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.3.7
langchain==0.0.350
langchain-openai==0.0.2
pinecone-client==2.2.4
faiss-cpu==1.7.4
requests==2.31.0
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1
numpy==1.25.2
sentence-transformers==2.2.2
httpx==0.25.2
tenacity==8.2.3 